refactor(seahaven-package): update dependencies and restructure manifest parts
- Removed `serde_yaml` dependency from `Cargo.toml` and `Cargo.lock`.
- Updated `toml` dependency to include the `preserve_order` feature.
- Refactored manifest parts by renaming and reorganizing modules, specifically moving `services` to `targets` and updating related imports.
- Adjusted deserialization tests to utilize TOML format instead of YAML, enhancing consistency in data handling.

This update improves the organization of the manifest structure and streamlines dependency management.
