+++ seahaven-package/Cargo.toml
@@ -13,9 +13,8 @@ once_cell = "1.21.3"
 regress = "0.10.3"
 semver = "1.0"
 serde = { version = "1.0", features = ["derive"] }
-serde_yaml = "0.9.34"
 thiserror = "2.0"
-toml = "0.8.20"
+toml = { version = "0.8.20", features = ["preserve_order"] }
 tracing = "0.1.41"
 
 [dev-dependencies]
