+++ seahaven-package/src/lib.rs
@@ -1,7 +1,3 @@
 //! # Seahaven Package
 
-#[macro_use]
-extern crate serde;
-
 pub mod manifest;
-pub mod transcoding;
