+++ seahaven-package/src/manifest/parts/root.rs
@@ -1,8 +1,8 @@
 use serde::de::Error as _;
 
 use super::{
-    name::Name,
-    services::{InitContainer, Service},
+    meta::PackageMeta,
+    targets::{InitContainer, Service},
 };
 
 /// The manifest for a Seahaven package.
@@ -48,31 +48,4 @@ impl<'de> serde::de::Deserialize<'de> for Manifest {
             init: manifest.init,
         })
     }
-}
-
-/// The package meta information
-#[derive(Debug, Clone, serde::Deserialize, serde::Serialize)]
-pub struct PackageMeta {
-    /// The name of the package
-    ///
-    /// This is the name of the package as it will be referenced in the workspace.
-    pub name: Name,
-
-    /// The version of the package
-    ///
-    /// This is an optional field.
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub version: Option<String>,
-
-    /// The description of the package
-    ///
-    /// This is an optional field.
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub description: Option<String>,
-
-    /// The readme file for the package
-    ///
-    /// This is an optional field.
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub readme: Option<String>,
 }
